+++ apps/backend/src/modules/points-of-interest/libs/validation-schemas/validation-schemas.ts
@@ -1,5 +1,5 @@
 export {
-	pointOfInterestCreateValidationSchema,
 	pointOfInterestUpdateValidationSchema,
-	pointsOfInterestSearchQueryValidationSchema,
+	pointsOfInterestCreateValidationSchema,
+	pointsOfInterestQueryValidationSchema,
 } from "@smartscapes/shared";