+++ apps/frontend/src/libs/modules/map-client/libs/enums/map-event-type.enum.ts
@@ -1,0 +1,7 @@
+const MapEventType = {
+	CLICK: "click",
+	DRAG_END: "dragend",
+	LOAD: "load",
+} as const;
+
+export { MapEventType };