+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -1,15 +1,20 @@
 type IconName =
 	| "analytics"
 	| "arrowDown"
+	| "caretDown"
+	| "caretUp"
 	| "close"
 	| "dashboard"
+	| "edit"
 	| "error"
 	| "eye"
 	| "eyeOff"
+	| "logout"
 	| "map"
 	| "message"
 	| "route"
 	| "tag"
+	| "trash"
 	| "user";
 
 export { type IconName };